% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScoreOrdinal.R
\name{AutoScore_Ordinal_rank}
\alias{AutoScore_Ordinal_rank}
\title{AutoScore-Ordinal STEP (1): Generate variable ranking List by machine
  learning (AutoScore-Ordinal Module 1)}
\usage{
AutoScore_Ordinal_rank(train_set, ntree = 100)
}
\arguments{
\item{train_set}{A processed \code{data.frame} that contains data to be
analyzed, for training. MIssing values and outliers should be cleaned
first. The outcome's name should be "label". Going through
\code{Preprocess} for preprocessing data and \code{\link{split_data}} for
splitting data, before AutoScore-Ordinal implementation, if needed.}

\item{ntree}{Number of trees in random forest algorithm for variable ranking,
default:100}
}
\value{
Returns a vector containing the list of variables and its ranking
  generated by machine learning (random forest)
}
\description{
AutoScore-Ordinal STEP (1): Generate variable ranking List by machine
  learning (AutoScore-Ordinal Module 1)
}
\details{
The first step in the AutoScore-Ordinal framework is variable
  ranking. We use random forest (RF), an ensemble machine learning algorithm,
  to identify the top-ranking predictors for subsequent score generation.
  This step correspond to Module 1 in the AutoScore-Ordinal paper, which is
  modified from the AutoScore paper.
}
\examples{
# See AutoScore-Ordinal Guidebook for the whole 5-step workflow
\dontrun{
ranking <- AutoScore_Ordinal_rank(train_set, ntree=100)
}

}
\references{
\itemize{
 \item{Breiman, L. (2001), Random Forests, Machine Learning 45(1), 5-32}
 \item{Xie F, Chakraborty B, Ong MEH, Goldstein BA, Liu N, AutoScore: A
 Machine Learning-Based Automatic Clinical Score Generator and Its
 Application to Mortality Prediction Using Electronic Health Records, JMIR
 Med Inform 2020;8(10):e21798, doi: 10.2196/21798}
}
}
\seealso{

}
