% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScoreOrdinal.R
\name{AutoScore_Ordinal_testing}
\alias{AutoScore_Ordinal_testing}
\title{AutoScoreOrdinal STEP(v): Evaluate the final score with ROC analysis
  (AutoScore Module 6)}
\usage{
AutoScore_Ordinal_testing(
  test_set,
  final_variables,
  cut_vec,
  score_table,
  with_label = TRUE,
  n_boot = 100
)
}
\arguments{
\item{test_set}{A processed \code{data.frame} that contains data for testing
purpose. This \code{data.frame} should have same format as \code{train_set}
(same variable names and outcomes)}

\item{final_variables}{A vector containing the list of selected variables,
selected from Step(ii). See Guidebook}

\item{cut_vec}{Generated from STEP(iii) \code{AutoScore_weighting()}.Please
follow the guidebook}

\item{with_label}{Set to TRUE if there are labels in the test_set and
performance will be evaluated accordingly (Default:TRUE).}

\item{scoring_table}{The final scoring table after fine-tuning, generated
from STEP(iv) \code{AutoScore_fine_tuning}.Please follow the guidebook}

\item{threshold}{Score threshold for the ROC analysis to generate
sensitivity, specificity, etc. If set to "best", the optimal threshold will
be calculated (Default:"best").}
}
\value{
A data frame with predicted score and the outcome for downstream visualization.
}
\description{
Domain knowledge is essential in guiding risk model development.
  For continuous variables, the variable transformation is a data-driven
  process (based on "quantile", "kmeans" or "decision_tree). In this step,
  the automatically generated cutoff values for each continuous variable can
  be fine-tuned by combining, rounding, and adjusting according to the
  standard clinical norm.  Revised \code{cut_vec} will be input  with domain
  knowledge to update scoring table. User can choose any cut-off values/any
  number of categories. Then final Scoring table will be generated. See
  Guidebook.
}
\examples{
\dontrun{
pred_score <- AutoScore_Ordinal_testing(test_set, final_variables, cut_vec,
scoring_table)
}
}
\seealso{
\code{\link{AutoScore_Ordinal_rank}},
  \code{\link{AutoScore_Ordinal_parsimony}},
  \code{\link{AutoScore_Ordinal_weighting}},
  \code{\link{AutoScore_Ordinal_fine_tuning}}.
}
