% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScoreOrdinal.R
\name{check_data}
\alias{check_data}
\title{AutoScoreOrdinal function: Check whether the input dataset fulfil the
  requirement of the AutoScoreOrdinal}
\usage{
check_data(data)
}
\arguments{
\item{data}{The data to be checked}
}
\value{
No return value, the result of the checking will be printed out.
}
\description{
AutoScoreOrdinal function: Check whether the input dataset fulfil the
  requirement of the AutoScoreOrdinal
}
\examples{
data("sample_data_ordinal")
check_data(sample_data_ordinal)
}
