% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_cutoffs}
\alias{compute_cutoffs}
\title{Compute the cutoff values for X%*%beta from theta terms of a CLM}
\usage{
compute_cutoffs(theta, z = NULL, link)
}
\arguments{
\item{theta}{theta terms of a CLM. For an outcome with J categories, there
should be J-1 elements in \code{theta}.}

\item{z}{A numeric vector for the predictor value. Make sure the range is
sufficient for all intersection between probability curves to intersect.
Default is \code{NULL} for \code{link = "logit"}.}

\item{link}{Link function of the CLM. Possible values are \code{"logit"}
(default), \code{"probit"}, \code{"cloglog"} and \code{"loglog"}.}
}
\value{
Returns a numeric vector of length J-1.
}
\description{
Compute the cutoff values for X%*%beta from theta terms of a CLM
}
