% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScoreOrdinal.R
\name{compute_mean_auc}
\alias{compute_mean_auc}
\title{Compute mean AUC for ordinal predictions}
\usage{
compute_mean_auc(y, fx)
}
\arguments{
\item{y}{An ordered factor representing the ordinal outcome, with length n
and J categories.}

\item{fx}{Either (i) a numeric vector of predictor (e.g., predicted scores)
of length n or (ii) a numeric matrix of predicted cumulative probabilities
with n rows and (J-1) columns.}
}
\value{
The mean AUC of J-1 cumulative AUCs (i.e., when evaluating
  the prediction of Y<=j, j=1,...,J-1).
}
\description{
Compute mean AUC for ordinal predictions
}
