% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_score_table}
\alias{compute_score_table}
\title{Compute scoring table}
\usage{
compute_score_table(train_set_2, validation_set_2, max_score, variable_list)
}
\arguments{
\item{max_score}{Restrict the largest possible score to avoid large partial
scores that are difficult to use in practice. Set to \code{NULL} to not
pose any restriction.}
}
\description{
Compute scoring table
}
