% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScoreOrdinal.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate model performance on data}
\usage{
evaluate_model(
  data,
  final_variables,
  score_table,
  cut_vec,
  n_boot,
  report_cindex = TRUE
)
}
\arguments{
\item{data}{A processed \code{data.frame} that contains data for validation
or testing purpose. This \code{data.frame} must have variable \code{label}
and should have same format as \code{train_set} (same variable names and
outcomes)}

\item{final_variables}{A vector containing the list of selected variables}

\item{cut_vec}{Generated from STEP(iii)
\code{AutoScore_Ordinal_weighting()}.Please follow the guidebook}

\item{report_cindex}{If generalised c-index should be reported alongside mean
AUC. Default is FALSE.}

\item{scoring_table}{The scoring table}
}
\value{
Returns a list of the mean AUC (mauc) and generalised c-index
  (cindex, if requested for) and their 95% bootstrap CIs (mauc_ci and cindex_ci).
}
\description{
Evaluate model performance on data
}
