% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{optimise_thresholds}
\alias{optimise_thresholds}
\title{Optimise thresholds for total score using cumulative ROC analyses}
\usage{
optimise_thresholds(y, total_score)
}
\arguments{
\item{y}{Observed ordinal outcome.}

\item{total_score}{Total score generated from AutoScore framework.}
}
\description{
Optimise thresholds for total score using cumulative ROC analyses
}
