% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_ordinal}
\alias{sample_data_ordinal}
\title{Simulated ED data with ordinal outcome}
\format{
An object of class \code{data.frame} with 20000 rows and 42 columns.
}
\usage{
sample_data_ordinal
}
\description{
Simulated data for 20,000 inpatient visits with demographic
  information and associated laboratory tests and vital signs measured in the
  emergency department (ED). Data were simulated based on the dataset
  analysed in the AutoScore-Ordinal paper. It is used for demonstration only
  in the Guidebook. Run \code{vignette("Guide_book", package = "AutoScore")}
  to see the guidebook or vignette.
}
\keyword{datasets}
