% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_mean_auc}
\alias{compute_mean_auc}
\title{Compute mean AUC for ordinal predictions}
\usage{
compute_mean_auc(y, fx)
}
\arguments{
\item{y}{An ordered factor representing the ordinal outcome, with length n
and J categories.}

\item{fx}{Either (i) a numeric vector of predictor (e.g., predicted scores)
of length n or (ii) a numeric matrix of predicted cumulative probabilities
with n rows and (J-1) columns.}
}
\value{
A \code{data.frame} with the cumulative AUCs (i.e., when evaluating
  the prediction of Y<=j, j=1,...,J-1) and 95% CI of AUC and the mean of
  cumulative AUCs. The SE of the mean AUC is approximated by assuming
  independence of cumulative AUCs.
}
\description{
Compute mean AUC for ordinal predictions
}
