% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScoreOrdinal.R
\name{compute_score_table}
\alias{compute_score_table}
\title{Internal function: Compute scoring table based on training dataset
  (AutoScore-Ordinal Module 3)}
\usage{
compute_score_table(train_set_2, validation_set_2, max_score, variable_list)
}
\arguments{
\item{train_set_2}{Processed training set after variable transformation
(AutoScore-Ordinal Module 2)}

\item{validation_set_2}{Processed validation set after variable
transformation (AutoScore-Ordinal Module 2)}

\item{max_score}{Maximum total score}

\item{variable_list}{List of included variables}
}
\value{
A scoring table
}
\description{
Compute scoring table based on training dataset
}
